Split up remotedb files
