+++ remotedb/remotedb.go
@@ -128,155 +128,3 @@ func (rd *RemoteDB) Iterator(start, end []byte) (db.Iterator, error) {
 	}
 	return makeIterator(dic), nil
 }
-
-func makeIterator(dic protodb.DB_IteratorClient) db.Iterator {
-	return &iterator{dic: dic}
-}
-
-func makeReverseIterator(dric protodb.DB_ReverseIteratorClient) db.Iterator {
-	return &reverseIterator{dric: dric}
-}
-
-type reverseIterator struct {
-	dric protodb.DB_ReverseIteratorClient
-	cur  *protodb.Iterator
-}
-
-var _ db.Iterator = (*iterator)(nil)
-
-func (rItr *reverseIterator) Valid() bool {
-	return rItr.cur != nil && rItr.cur.Valid
-}
-
-func (rItr *reverseIterator) Domain() (start, end []byte) {
-	if rItr.cur == nil || rItr.cur.Domain == nil {
-		return nil, nil
-	}
-	return rItr.cur.Domain.Start, rItr.cur.Domain.End
-}
-
-// Next advances the current reverseIterator
-func (rItr *reverseIterator) Next() {
-	var err error
-	rItr.cur, err = rItr.dric.Recv()
-	if err != nil {
-		panic(fmt.Sprintf("RemoteDB.ReverseIterator.Next error: %v", err))
-	}
-}
-
-func (rItr *reverseIterator) Key() []byte {
-	if rItr.cur == nil {
-		panic("key does not exist")
-	}
-	return rItr.cur.Key
-}
-
-func (rItr *reverseIterator) Value() []byte {
-	if rItr.cur == nil {
-		panic("key does not exist")
-	}
-	return rItr.cur.Value
-}
-
-func (rItr *reverseIterator) Error() error {
-	return nil
-}
-
-func (rItr *reverseIterator) Close() {}
-
-// iterator implements the db.Iterator by retrieving
-// streamed iterators from the remote backend as
-// needed. It is NOT safe for concurrent usage,
-// matching the behavior of other iterators.
-type iterator struct {
-	dic protodb.DB_IteratorClient
-	cur *protodb.Iterator
-}
-
-var _ db.Iterator = (*iterator)(nil)
-
-func (itr *iterator) Valid() bool {
-	return itr.cur != nil && itr.cur.Valid
-}
-
-func (itr *iterator) Domain() (start, end []byte) {
-	if itr.cur == nil || itr.cur.Domain == nil {
-		return nil, nil
-	}
-	return itr.cur.Domain.Start, itr.cur.Domain.End
-}
-
-// Next advances the current iterator
-func (itr *iterator) Next() {
-	var err error
-	itr.cur, err = itr.dic.Recv()
-	if err != nil {
-		panic(fmt.Sprintf("remoteDB.Iterator.Next error: %v", err))
-	}
-}
-
-func (itr *iterator) Key() []byte {
-	if itr.cur == nil {
-		return nil
-	}
-	return itr.cur.Key
-}
-
-func (itr *iterator) Value() []byte {
-	if itr.cur == nil {
-		panic("current poisition is not valid")
-	}
-	return itr.cur.Value
-}
-
-func (itr *iterator) Error() error {
-	return nil
-}
-
-func (itr *iterator) Close() {
-	err := itr.dic.CloseSend()
-	if err != nil {
-		panic(fmt.Sprintf("Error closing iterator: %v", err))
-	}
-}
-
-type batch struct {
-	db  *RemoteDB
-	ops []*protodb.Operation
-}
-
-var _ db.Batch = (*batch)(nil)
-
-func (bat *batch) Set(key, value []byte) {
-	op := &protodb.Operation{
-		Entity: &protodb.Entity{Key: key, Value: value},
-		Type:   protodb.Operation_SET,
-	}
-	bat.ops = append(bat.ops, op)
-}
-
-func (bat *batch) Delete(key []byte) {
-	op := &protodb.Operation{
-		Entity: &protodb.Entity{Key: key},
-		Type:   protodb.Operation_DELETE,
-	}
-	bat.ops = append(bat.ops, op)
-}
-
-func (bat *batch) Write() error {
-	if _, err := bat.db.dc.BatchWrite(bat.db.ctx, &protodb.Batch{Ops: bat.ops}); err != nil {
-		return errors.Errorf("remoteDB.BatchWrite: %v", err)
-	}
-	return nil
-}
-
-func (bat *batch) WriteSync() error {
-	if _, err := bat.db.dc.BatchWriteSync(bat.db.ctx, &protodb.Batch{Ops: bat.ops}); err != nil {
-		return errors.Errorf("RemoteDB.BatchWriteSync: %v", err)
-	}
-	return nil
-}
-
-func (bat *batch) Close() {
-	bat.ops = nil
-}
